@@ -10,20 +10,20 @@
 template <class IndexerFactoryFunc, class QueryFunc, class ListFactoryFunc>
 void benchmark(
     int qn, 
-    int d, 
     const float** query, 
     const Result** results,
-    const std::string& outputFilename,  
+    // const std::string& outputFilename,  
+    FILE* fp, 
     const IndexerFactoryFunc& fIndexerFactory, 
     const QueryFunc& fQuery, 
     const ListFactoryFunc& fListFactory)
 {
-    std::cout << outputFilename << std::endl;
-    FILE* fp = fopen(outputFilename.c_str(), "a+");
-    if (!fp) {
-        printf("cannot open file %s\n", outputFilename.c_str());
-        return;
-    }
+    // std::cout << outputFilename << std::endl;
+    // FILE* fp = fopen(outputFilename.c_str(), "a+");
+    // if (!fp) {
+    //     printf("cannot open file %s\n", outputFilename.c_str());
+    //     return;
+    // }
 
 	timeval start_time, end_time;
     gettimeofday(&start_time, NULL);
@@ -34,13 +34,15 @@ void benchmark(
     printf("Indexing Time: %f Seconds\n\n", indexing_time);
     fprintf(fp, "Indexing Time: %f Seconds\n\n", indexing_time);
 
-    int kMIPs[] = { 1, 2, 5, 10, 100 };
+    // int kMIPs[] = { 1, 2, 5, 10, 100 };
+    int kMIPs[] = { 1, 2, 5, 10};
     int max_round = sizeof(kMIPs) / sizeof(int);
     int top_k = -1;
 
     float percentage = 0.5;
     // printf("Top-k weighted ANN of AWS_SPHERE_LSH: \n");
-    printf("  Top-k\t\t%.2fRatio\tTime (ms)\tRecall\n", percentage);
+    // printf("  Top-k\t\t%.2fRatio\tTime (ms)\tRecall\n", percentage);
+    printf("  Top-k\t\tRatio\tTime (ms)\tRecall\n");
     for (int num = 0; num < max_round; num++) {
         gettimeofday(&start_time, NULL);
 
@@ -59,6 +61,7 @@ void benchmark(
             recall += calc_recall(top_k, (const Result*)results[i], list.get());
 
             ratios[i] = calc_ratio(top_k, (const Result*)results[i], list.get());
+            overall_ratio += ratios[i];
             // printf("%d, |res|=%d, [%d, %f]; [%d, %f]; %f, %f\n", i, list->size(), list->ith_id(0), list->ith_key(0), results[i][0].id_, results[i][0].key_, recalls[i], ratios[i]);
         }
 
@@ -75,41 +78,41 @@ void benchmark(
         overall_ratio = overall_ratio / qn;
         recall = recall / qn;
 
-        printf("  %3d\t\t%.4f\t\t%.4f\t\t%.2f%%\n", top_k, medianRatio,
+        // printf("  %3d\t\t%.4f\t\t%.4f\t\t%.2f%%\n", top_k, medianRatio,
+        //     avg_runtime, recall);
+        printf("  %3d\t\t%.4f\t\t%.4f\t\t%.2f%%\n", top_k, overall_ratio,
             avg_runtime, recall);
-        for (int i = 0; i < qn; i++) {
-            fprintf(fp, "%d-%f\n", i, ratios[i]);
-        }
+        // for (int i = 0; i < qn; i++) {
+        //     fprintf(fp, "%d-%f\n", i, ratios[i]);
+        // }
         fprintf(fp, "%d\t%f\t%f\t%f\n", top_k, medianRatio, avg_runtime, recall);
     }
     printf("\n");
     fprintf(fp, "\n");
-    fclose(fp);
+    // fclose(fp);
     return;
 }
 
-//with weight
 //fif::()->unique_ptr<Indexer>
 //lff::int top_k->unique_ptr<ListType>
-//qf::indexer&->int k->double** query->double** weight->ListType*->IO
+//qf::indexer&->int k->double** query->ListType*->IO
 template <class IndexerFactoryFunc, class QueryFunc, class ListFactoryFunc>
-void benchmark(
+void benchmark_multiplek(
     int qn, 
-    int d, 
     const float** query, 
-    const float** weight,
-    const Result** results, 
-    const std::string& outputFilename, 
+    const Result** results,
+    std::vector<int>& check_ks,
+    FILE* fp, 
     const IndexerFactoryFunc& fIndexerFactory, 
     const QueryFunc& fQuery, 
     const ListFactoryFunc& fListFactory)
 {
-    std::cout << outputFilename << std::endl;
-    FILE* fp = fopen(outputFilename.c_str(), "a+");
-    if (!fp) {
-        printf("cannot open file %s\n", outputFilename.c_str());
-        return;
-    }
+    // std::cout << outputFilename << std::endl;
+    // FILE* fp = fopen(outputFilename.c_str(), "a+");
+    // if (!fp) {
+    //     printf("cannot open file %s\n", outputFilename.c_str());
+    //     return;
+    // }
 
 	timeval start_time, end_time;
     gettimeofday(&start_time, NULL);
@@ -120,108 +123,136 @@ void benchmark(
     printf("Indexing Time: %f Seconds\n\n", indexing_time);
     fprintf(fp, "Indexing Time: %f Seconds\n\n", indexing_time);
 
-    int kMIPs[] = { 1, 2, 5, 10, 100};
+    printf("memory-usage: %ld (bytes)\n", lsh->get_memory_usage());
+    fprintf(fp, "memory-usage: %ld (bytes)\n", lsh->get_memory_usage());
+
+    // int kMIPs[] = { 1, 2, 5, 10, 100 };
+    int kMIPs[] = { 1, 2, 5, 10};
     int max_round = sizeof(kMIPs) / sizeof(int);
     int top_k = -1;
 
-    float percentage = 0.5;
-    // printf("Top-k weighted ANN of AWS_SPHERE_LSH: \n");
-    printf("  Top-k\t\t%.2fRatio\tTime (ms)\tRecall\n", percentage);
-    for (int num = 0; num < max_round; num++) {
-        gettimeofday(&start_time, NULL);
 
-        top_k = kMIPs[num];
-        auto list = fListFactory(top_k);
-        std::vector<float> ratios(qn);
-        std::vector<float> recalls(qn);
+    for(int check_k:check_ks){    
+        float percentage = 0.5;
 
-        float overall_ratio = 0.0f;
-        float recall = 0.0f;
-        for (int i = 0; i < qn; ++i) {
-            list->reset();
-            fQuery(*lsh, top_k, query[i], weight[i], list.get());
+        printf("check_k=%d\n", check_k);
+        fprintf(fp, "check_k=%d\n", check_k);
 
-            // printf("%d, %d\n", num, i);
+        // printf("Top-k weighted ANN of AWS_SPHERE_LSH: \n");
+        // printf("  Top-k\t\t%.2fRatio\tTime (ms)\tRecall\n", percentage);
+        printf("  Top-k\t\tRatio\t\tTime (ms)\tRecall\n");
+        for (int num = 0; num < max_round; num++) {
+            gettimeofday(&start_time, NULL);
 
-            recalls[i] = calc_recall(top_k, (const Result*)results[i], list.get());
-            recall += calc_recall(top_k, (const Result*)results[i], list.get());
+            top_k = kMIPs[num];
+            auto list = fListFactory(top_k);
+            std::vector<float> ratios(qn);
+            std::vector<float> recalls(qn);
 
-            ratios[i] = calc_ratio(top_k, (const Result*)results[i], list.get());
-            // printf("%d, [%d, %f]; [%d, %f]; %f, %f\n", i, list->ith_id(0), list->ith_key(0), results[i][0].id_, results[i][0].key_, recalls[i], ratios[i]);
+            float overall_ratio = 0.0f;
+            float recall = 0.0f;
+            for (int i = 0; i < qn; ++i) {
+                list->reset();
+                fQuery(*lsh, top_k, check_k, query[i], list.get());
+
+                recalls[i] = calc_recall(top_k, (const Result*)results[i], list.get());
+                recall += calc_recall(top_k, (const Result*)results[i], list.get());
+
+                ratios[i] = calc_ratio(top_k, (const Result*)results[i], list.get());
+                overall_ratio += ratios[i];
+                // printf("%d, |res|=%d, [%d, %f]; [%d, %f]; %f, %f\n", i, list->size(), list->ith_id(0), list->ith_key(0), results[i][0].id_, results[i][0].key_, recalls[i], ratios[i]);
+            }
+
+            gettimeofday(&end_time, NULL);
+            float runtime = end_time.tv_sec - start_time.tv_sec + (end_time.tv_usec -
+                start_time.tv_usec) / 1000000.0f;
+            float avg_runtime = (runtime * 1000.0f) / qn;
+
+            sort(ratios.begin(), ratios.end());
+            // sort(recalls.begin(), recalls.end());
+            float medianRatio = ratios[qn * percentage];
+            // float medianRecall = recalls[qn*percentage];
+
+            overall_ratio = overall_ratio / qn;
+            recall = recall / qn;
+
+            // printf("  %3d\t\t%.4f\t\t%.4f\t\t%.2f%%\n", top_k, medianRatio,
+            //     avg_runtime, recall);
+            printf("  %3d\t\t%.4f\t\t%.4f\t\t%.2f%%\n", top_k, overall_ratio,
+                avg_runtime, recall);
+            // for (int i = 0; i < qn; i++) {
+            //     fprintf(fp, "%d-%f\n", i, ratios[i]);
+            // }
+            fprintf(fp, "%d\t%f\t%f\t%f\n", top_k, medianRatio, avg_runtime, recall);
         }
-
-        gettimeofday(&end_time, NULL);
-		float runtime = end_time.tv_sec - start_time.tv_sec + (end_time.tv_usec -
-			start_time.tv_usec) / 1000000.0f;
-        float avg_runtime = (runtime * 1000.0f) / qn;
-
-        sort(ratios.begin(), ratios.end());
-        // sort(recalls.begin(), recalls.end());
-        float medianRatio = ratios[qn * percentage];
-        // float medianRecall = recalls[qn*percentage];
-
-        overall_ratio = overall_ratio / qn;
-        recall = recall / qn;
-
-        printf("  %3d\t\t%.4f\t\t%.4f\t\t%.2f%%\n", top_k, medianRatio,
-            avg_runtime, recall);
-        for (int i = 0; i < qn; i++) {
-            fprintf(fp, "%d-%f\n", i, ratios[i]);
-        }
-        fprintf(fp, "%d\t%f\t%f\t%f\n", top_k, medianRatio, avg_runtime, recall);
+        printf("\n");
+        fprintf(fp, "\n");
     }
-    printf("\n");
-    fprintf(fp, "\n");
-    fclose(fp);
+
+    // fclose(fp);
     return;
 }
 
 template <class IndexerFactoryFunc, class QueryFunc>
 void benchmarkMinklist(
     int qn,
-    int d,
     const float** query,
-    const float** weight,
     const Result** results,
-    const std::string& outputFilename, 
+    FILE* fp, 
     const IndexerFactoryFunc& fIndexerFactory, 
     const QueryFunc& fQuery)
 {
     const auto& fListFactory = [&](int top_k){
         return make_unique<MinK_List>(top_k);
     };
-    benchmark(qn, d, query, weight, results, outputFilename, fIndexerFactory, fQuery, fListFactory);
-}
-
-//without weight
-template <class IndexerFactoryFunc, class QueryFunc>
-void benchmarkMinklist(
-    int qn,
-    int d,
-    const float** query,
-    const Result** results,
-    const std::string& outputFilename, 
-    const IndexerFactoryFunc& fIndexerFactory, 
-    const QueryFunc& fQuery)
-{
-    const auto& fListFactory = [&](int top_k){
-        return make_unique<MinK_List>(top_k);
-    };
-    benchmark(qn, d, query, results, outputFilename, fIndexerFactory, fQuery, fListFactory);
+    // benchmark(qn, query, results, outputFilename, fIndexerFactory, fQuery, fListFactory);
+    benchmark(qn, query, results, fp, fIndexerFactory, fQuery, fListFactory);
 }
 
 template <class IndexerFactoryFunc, class QueryFunc>
 void benchmarkMaxklist(
     int qn,
-    int d,
     const float** query,
     const Result** results,
-    const std::string& outputFilename, 
+    FILE* fp, 
     const IndexerFactoryFunc& fIndexerFactory, 
     const QueryFunc& fQuery)
 {
     const auto& fListFactory = [&](int top_k){
         return make_unique<MaxK_List>(top_k);
     };
-    benchmark(qn, d, query, results, outputFilename, fIndexerFactory, fQuery, fListFactory);
+    benchmark(qn, query, results, fp, fIndexerFactory, fQuery, fListFactory);
+}
+
+template <class IndexerFactoryFunc, class QueryFunc>
+void benchmarkMinklist(
+    int qn,
+    const float** query,
+    const Result** results,
+    std::vector<int>& check_ks,
+    FILE* fp, 
+    const IndexerFactoryFunc& fIndexerFactory, 
+    const QueryFunc& fQuery)
+{
+    const auto& fListFactory = [&](int top_k){
+        return make_unique<MinK_List>(top_k);
+    };
+    // benchmark(qn, query, results, outputFilename, fIndexerFactory, fQuery, fListFactory);
+    benchmark_multiplek(qn, query, results, check_ks, fp, fIndexerFactory, fQuery, fListFactory);
+}
+
+template <class IndexerFactoryFunc, class QueryFunc>
+void benchmarkMaxklist(
+    int qn,
+    const float** query,
+    const Result** results,
+    std::vector<int>& check_ks,
+    FILE* fp, 
+    const IndexerFactoryFunc& fIndexerFactory, 
+    const QueryFunc& fQuery)
+{
+    const auto& fListFactory = [&](int top_k){
+        return make_unique<MaxK_List>(top_k);
+    };
+    benchmark_multiplek(qn, query, results, check_ks, fp, fIndexerFactory, fQuery, fListFactory);
 }
