@@ -3,6 +3,7 @@
 #include <string>
 #include <chrono>
 #include <iostream>
+#include <stack>
 
 class MyTimer
 {
@@ -23,13 +24,19 @@ public:
 	}
 	void end() {
 		if(isRunning){
-			auto endTime = std::chrono::system_clock::now();
+			endTime = std::chrono::system_clock::now();
 			t_type t = endTime - startTime;
-			_tmMap()[name] += t;
-			_tmMapCnt()[name]++;
+			// printf(" %s: %f\n", name.c_str(), t.count());
+			if(name != ""){
+				_tmMap()[name] += t;
+				_tmMapCnt()[name]++;
+			}
 			isRunning = false;
 		}
 	}
+	double getTime(){
+		return (endTime-startTime).count();
+	} 
 
 	template<typename F, typename... Args>
 	static t_type funcTime(F&& func, Args&&... args){
@@ -39,9 +46,20 @@ public:
 		return t_type(t1-t0);
 	}
 
+	static void pusht() {
+		auto t = std::chrono::system_clock::now();
+		return get_tstack().push(t);
+	}
+	static double popt() {
+		auto t = get_tstack().top();
+		get_tstack().pop();
+        auto duration = std::chrono::duration_cast<t_type>(std::chrono::system_clock::now() - t);
+		return duration.count();
+	}
 
 	static void clear() {
 		_tmMap().clear();
+		_tmMapCnt().clear();
 	}
 	static void clear(const std::string &name) {
 		_tmMap()[name] = t_type::zero();
@@ -49,6 +67,14 @@ public:
 	static double get(const std::string &name) {
 		return _tmMap()[name].count();
 	}
+
+	static std::stack< std::chrono::time_point<std::chrono::system_clock> >& get_tstack()
+	{
+		static std::stack< std::chrono::time_point<std::chrono::system_clock> > _tstack;
+		return _tstack;
+	}
+	
+
 //	static double get(const char *name) {
 //		//std::string n(name);
 //		return g_tmMap[std::string(name)].count();
@@ -63,12 +89,13 @@ public:
 //		outp << "Time for " << name << ": " << get(name) << "  avg:" << get(name) / cnt(name) << std::endl;
 //	}
 	static void print(const std::string &name, std::ostream &outp = std::cout) {
-		outp << "Time for " << name << ": " << get(name) << "  avg:" << get(name) / cnt(name) << std::endl;
+		outp << "Time for " << name << ": " << get(name) << "  avg:" << get(name) / cnt(name) << "  cnt:" << cnt(name) << std::endl;
 	}
 	static void printAll(std::ostream &outp = std::cout) {
 		for (auto it = _tmMap().begin(); it != _tmMap().end(); it++) {
 			print(it->first, outp);
 		}
+		clear();
 	}
 	static void printCurTime(std::ostream &outp = std::cout){
 		auto time_point = std::chrono::system_clock::now();
@@ -99,6 +126,8 @@ private:
 //	static std::map<std::string, t_type> g_tmMap;
 //	static std::map<std::string, int> g_tmMapCnt;
 	std::chrono::time_point<std::chrono::system_clock> startTime;
+	std::chrono::time_point<std::chrono::system_clock> endTime;
 	std::string name;
 	bool isRunning;
+
 };
